Added bash_local auto-creation concept
I'm trying to cut down on the chance of error when changing bashrc
settings. I don't want to accidentally push stuff I shouldn't, so if I
leave a nice big note to myself to use the bash_local file and I also
have it automatically created and explicitly not tracked (i.e. no
symbolic link) it should be hard for me to screw it up.